mappings: rename wikidata to entities
as it's used by both inv entities and wd entities
