+++ server/db/elasticsearch/mappings/entities.js
